requirements.txt
nicegui
pyyaml
python-dotenv
requests
